[INTERNAL][FIX] sap.ui.rta: fix stretching in SimpleForm
Use overlay's geometry instead of overlay.getElement().$()

After a change to a field the stretching gets re-evaluated.
But due to the structure of the SimpleForm at this point the original control
is not available via the getElement function of the overlay,
but the overlay's geometry still has the correct reference.
There is still some flickering, but that is because the SimpleForm uses non-controls,
so we can only add the stretching after the object is created.

Also did some code clean-up

BCP: 2180240965
Change-Id: I9fefbab8b2c4bf0a2bb8841e9370c62348ab6a93