[FEATURE] sap.m.Button: screen reader announses text, icon or tooltip changes
When the button gets focused, an aria-live HTML attribute with value
"polite" is included. If there is a change on the text, icon or tooltip
of the button, this attribute makes the screen reader to announce the
new value. On focus out the value of the aria-live attribute is changed
to "off", as announcements are not wanted when the button is now focused.

JIRA: BGSOFUIBALKAN-5769
Change-Id: I9108ead2fa275a23ea1b1b5d006da4b4aaf6f423