[INTERNAL][FIX]sap.ui.rta: Prevent move of direct children in template aggregations
Elements which are direct children of a template aggregation should not be
movable since their position is defined by the data which they
are bound to. Being able to move one list item in e.g. a bound list
from one position to another in the template aggregation is not a
valid change - only the underlying data should define this position.
The binding context check during the movable check for such elements
is also causing general problems as in the linked incident.

This change checks if an element is a direct child of a template
during the movable evaluation of the RTAElementMover, returning false
false in this case and not proceeding with the binding context check.

Change-Id: I42d8e68e9da9857a119ddd319b72f95ac195d8d0
BCP: 2170173923