[INTERNAL] Demo Kit: Removed third party google-code-prettify library
The third party google-code-prettify library in no longer used in the
Demo Kit application.

Instead of it, we are now using the highlight.js third party library.

Additionally a small bug is fixed where "pre" tags are not found in
the SubApiDetail controller file when they are supposed to be processed
by the highlight.js library.

JIRA: BGSOFUIPIRIN-5400
Change-Id: I07e469c07cc1c35bf0bfe7bfb0308275365c21b8