[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Fixed line spacing
Problem: Visual tests fail, due to a recent change in ObjectStatus,
which changes the vertical alignment of texts, as a side effect -
changes also the line spacing in OPL's header facets.

Solution: The line-height is adjusted, so that the new vertical
alignment, does not influence the line-height.

BCP: N/A
Change-Id: I62b1ad52abd97cd73d64166de7eb8c6319797ca3