[FIX]sap.ui.fl: Condenser changes reorder fix
If the condenser calculated a different order for changes with
status "updated", they were not being marked as "reorder" by
the storage and being persisted in the wrong order.

This fix ensures that the "update" changes are also evaluated
for reorder.

Change-Id: Ic62fb0252a19fc0a6beafd7aa60db866d83eeec5
(cherry picked from commit dcacfa31fbbd37692bee9eec509ffdfaf807ff26)
BCP: 2170211539
CR-Id: 002075125800002544072021