[INTERNAL] sap.ui.fl removes the getPendingAction
The getPendingAction of Changes and Variants was only a wrapper of the getState function. Thus, all occurrences were replaced with the getState function.

In addition the states were extracted to the sap/ui/fl/apply/_internal/flexObjects/States. These are referenced in general and only the sap.ui.fl.Changes.states used in case it is clearly referencing a Change.

Change-Id: I5d05ade2c69003b92c34a65b439a704a71754c13