[INTERNAL] sap.ui.fl...Loader does not change IDs on cloning
In case cloned changes (OVP-id migration) the condensing logic only detects the change IDs of the clones and sends these as a delete request to the back end. To ensure the deletion on back end side finds the correct IDs the clones mut have the same ID as its origins.

Change-Id: Ie5f12aeaa688c12cbee73bf62f814306f56c95ab