[FIX] sap.m.SearchField: Updates on suggestion items are now shown
Problem: Suggestion items' that are bound to ODataV4 model appear blank
in the suggestions list.

Root cause: The properties are set late(asynchronously), after the suggestions list is opened.
When any property of sap.m.SuggestionItem is set, the invalidation flow is suppressed.

Solution: Do not suppress invalidation.

Change-Id: I1a62fd31b744b3d4f3c40215f312ea9276f6d416
BCP: 2170217899
Snippix: 84252