[FIX] Demo Kit: Fix autoFocus logic in FCL sample
Problem: In one of the samples of FlexibleColumnLayout, the autofocusing
logics is not working for the middle column.

Solution: The autofocusing works only for the first rendering of each
NavContainer. In this sample, however, it did not work for the
NavContainer in the middle column, due to a wrong routes configuration
in the manifest.json. As for the "master" view, the first and the middle
columns were reffered, the NavContainer in the middle got rendered
before the user navigates to it. That's why, after navigation, the
autoFocus is not applied, because this was actually a second rendering of
it.

BCP: 2170162405
Change-Id: I6daa8d8e36702ee625398bf7ee3c42137165d3e2