[FIX] sap.m.OverflowToolbarButton: Updated OTB configuration
Problem: The OverflowToolbarButton doesn't autoclose the OverflowToolbar
Popover unless explicitly instructed with press event by the App developer.

Root Cause: Missing information in the configuration object which is
returned as OverflowToolbar oConfig by the OverflowToolbarButton.

Solution: Updated the OverflowToolbar oConfig of the OverflowToolbarButton
with the correct properties. Now the OverflowToolbarButton automatically
closes the OverflowToolbar Popover as the ordinary Button does.

BCP: 2170208591
Change-Id: I61531c5261db6595058b863c33abc0198a30025f