[FIX] Demo Kit: Fixed incorrect link redirect to demo applications
Problem: When we try to open some of the applications in the Demo Apps
section of the Demo Kit, we end up with one additional unnecessary
hash ('#') in the URL which prevents us to open the desired application.
The newly opened web page is not found because the URL is incorrect.

Root Cause: In the code we have a check which is supposed to prevent the
addition of the hash in case of a standalone pages (including .html), but
we check if the URL ends up with ".html", instead of checking if ".html"
exists in the URL itself. In this particular case we open an ".html" page
but since we have a search query for the application's theme after the
".html" part, our prevention check fails.

Solution: With a certain regular expression rules, we check if there
is the ".html" substring in the URL as well as if there are additional
hashes or search queries.

BCP: 2170199589
CR-Id: 002075125800002789862021
Change-Id: Ide6abd8c1f321bab1a692be298b58bfb7bcb82c6
(cherry picked from commit b86a1b9c86911971ef53516cb6efabec01c05ac9)