[FIX] sap.m.Select: Correct focus on combi devices
Problem:
Focus is lost when the select is focused via touch on a touch-enabled
screen of a combi device

Findings:
After investigation we see that the browser removes the focus on 'tap' if
the focusable element has "position: absolute" or is under an element that
has "position: absolute".

Solution:
A solution of refactoring the DOM structure (to remove "position: absolute")
is not considered for now due to restructions for ACC (the DOM structure was
already previously reworked to the current one to cover ACC requirements)

A second solution that ensures the select *wrapper* is focusable with touch
is proposed. This second solution was is place earlier and preventend the
current issue before change Ib1390a6a8026615197280f81637528f1b47819eb
Therefore an alternative solution for BCP 2180201197 is proposed here:

Problem of BCP 2180201197:
The OverflowToolbar.prototype._applyFocus sets the focus to the *wrapper*
of the Select, rather than to the select's inner focuable element (returned from
Select.prototype.getFocusDomRef)
Alternative solution for BCP 2180201197:
Ensure that the OverflowToolbar.prototype._applyFocus re-applies the focus
using Element.prototype.focus which respects the dom ref returned from
Element.prototype.getFocusDomRef

BCP: 002075129500006164082021
BCP: 2180201197
Change-Id: I29bdb157076f0e022264c27e367dba3d0609371a