[INTERNAL] sap.m.Select: Tooltip added for dropdown arrow
Problem: When there is tooltip set to the sap.m.Select, only when we
hover over the its selected item we get a tooltip. The reporter claims
that the tooltip should be visible if we hover the dropdown arrow too.

Solution: We make the desired tooltip visible when we hover over the
dropdown arrow as well.

BCP: 2170171593
Change-Id: If1834eab857dc302bb4fb5528fb46b95b4da4fbb