[INTERNAL] sap.uxap.ObjectPageLayout: remove delay on contentSize change
Background: After content size is changed, the ObjectPageLayout is doing
recalculations whether scroll bar is needed or not. This recalculations
are done with a delay, which slows down performance measurements.

Solution: Now the recalculations for the scroll bar are done immediately
after the content size change handler is called, without no delay.

BCP: 2180206643
Change-Id: Ieec87c32c8da05f723ff9b144097ea3cb85318e1