[FIX] sap.uxap.ObjectPageLayout: Removed single section upper line
Since change 5287638 there is a visual problem with ObjectPageLayout
sections container, that the separation line between sections, above
the title of the section is visible when there is only one section and
no Anchor Bar is presented. When the anchor bar is visible, the first
secion title is invisible and the Anchor Bar link is visually used
as a title.

BCP: 2170230576
Change-Id: I1bfc15ac091188432ec354daaeed0783da116d22