[INTERNAL] sap.ui.integration: Allow customization of property editor default configs
Currently it is possible to define default configurations for property
editors via the configMetadata and then customize them for single props.
However, if an existing editor generally needs to be customized or if a
property supports different editor types via bindings, it is not possible
to do so on a base editor level.
This commit adds a further layer for such configs so that simple config
changes don't have to be done by overriding the editors.

Change-Id: I974a519867409e516fffea72b2c1f60c95aed355
JIRA: TEAM42-142