[INTERNAL] sap.f.FlexibleColumnLayout: Fix autoFocus logic in ACC sample
Problem: In one of the samples of FlexibleColumnLayout, the autofocusing
logics is not working for the middle column.

Solution: The autofocusing works only for the first rendering of each
NavContainer. In this sample, however, it did not work for the
NavContainer in the middle column, due to a wrong routes configuration
in the manifest.json. As for the "master" view, the first and the middle
columns were reffered, the NavContainer in the middle got rendered
before the user navigates to it. That's why, after navigation, the
autoFocus is not applied, because this was actually a second rendering of
it.

Change-Id: Ie981f693ddaddd546b287f86f1cb8736c790c479
BCP: N/A