[FIX] NumberFormat: Currency symbols should not be affected by trailingCurrencyCode option
Try to detect whether a given measure is actually a currency code. If
not, do not use the "trailingCurrencyCode" pattern for formatting.

If format option "customCurrencies" is supplied, a measure that
can't be found in the given list of currency codes is treated like a
symbol (or like a currency code with the "trailingCurrencyCode" option
set to false for that matter).

In all other cases, a given measure is matched against the CLDR data,
globally configured custom currencies and finally the three-letter
alphabetic code pattern as defined in the ISO 4217 standard.

Change-Id: Id1b021c7f5a2565b3de9ca96bbb515e82a88573f
CR-Id: 002075125800003249542021
BCP: 002075129500006756752021
JIRA: CPOUI5FOUNDATION-399