[FEATURE] sap.m.ComboBox: Clear icon implementation
This implementation introduces new public property called
'showClearIcon'.
Setting the property to true will make an "x" icon appear
in the end of the input part of the ComboBox control, which
when pressed will remove the typed in or already set value.

Two new demokit samples were added - one for each control
where the clear icon feature is switched on.

JIRA: BGSOFUIRILA-3195
Change-Id: Icd89353e49b47395e36d083884e39903bea7d571