[INTERNAL] Core: load library-preload-lazy.js for some lazy dependencies
To avoid the sync library loading in the head of some sap.m modules,
lazy dependencies are used to load bundle information for sap.ui.unified
and sap.ui.layout upfront. When later a module is required from those
libraries, then the loader will load the corresponding bundle.

Note: this only improves the situation regarding AMD dependencies. All
other functionality of a Core#loadLibrary call is still missed that way
(e.g preloading of texts, CSS will only be loaded when initLibrary is
called for the library).

To state it differently: an explicit, async Core#loadLibrary() call
right before modules from that library are required, is still the
preferred solution.

Change-Id: Ic3e96d00eca3151d2a3289cb766df96a9d559d05