[FIX] P13nDialog: IconTabBar items border fix
Issue: When the app is in Compact mode  the IconTabBar items are
focused, then we are facing missing top border side.

Solution: Add css class which is pointing specificly the Compact scenario
and only for ItemTabBar items in P13nDialog, which will increase the
padding-top and make the full border visible.

Change-Id: Ib71d1c2433155c30ba51825490ff69dd0031fb76
BCP: 2180308710