[FIX] sap.uxap.ObjectPageSubSection: Fix focus issue
Problem: The ShowMore and ShowLess buttons are always focused, when
pressed, and this behavior was implemented some time ago, after
complaints that the text of the button is getting changed, but the user
does not get notified for it, as the new text is not read out. However,
now some users want to move the focus to the newly shown blocks, for
example, if there is an editable field (Input), as otherwise, he may not
be aware what and where has just appeared on the screen.

Solution: As there is not always an editable field, shown in more
blocks, and also users have different preferences about what should be
focused and read out, now we provide them a way to move the focus in the
onAfterRendering logic of the BlockBase's controller. If the focus was
moved by intention from the application side - we do not restore it on
the ShowMore/ShowLess button. If it was not moved itentionally on a
different DOM element, because of the rerendering of the button, the
focus is moved on the document's body. In this case, we restore it to
the newly shown SeeMore/SeeLess button, keeping the existing logic.

BCP: 2170200164
Change-Id: Iec9f1181d198b6fa66074dfcbc55c89f6208e552