[INTERNAL] sap.m.Select: Change aria-roledescription to Listbox
The term "Combobox" means there are multiple ways to input a value, for example by typing it and choosing it.
With select the only way to input the value is by selecting it from a predefined options list. So base on user feedback the
value is changed to just "Listbox"

Change-Id: I623d4b1918fdc2008341e9c63f0c97b0cf05156f
BCP: 2180181715