[INTERNAL] sap.f.DynamicPage: Fixed background issue in FLP
Problem:
Visual issue in FLP when part of the FLP background shows bellow
the transparent area occupied by a header hidden with visibility: hidden
(as that area may still be visible to the user by being outside the scroll
overflow).

Findings:
Part of the area bellow hidden header made be outside the scroll
overflow for reasons outlined in JIRA: BGSOFUIPIRIN-4474

Solution:
Ensures that when the header is hidden with visibility: hidden
the area that it occupies still has the required background.

BCP: 2170246414
Change-Id: I424bc166b91c193b57d2089741d9b065c3a1f240