[INTERNAL][FIX] sap.ui.fl - getComponent check before getUShellService call
Recently the getUShellService call is switched to be async. To aviod
some async issues in some cases the services gets preloaded during
instatiation of an module and used inside. But this was not exactly
the behaviour before the ushell service call gets async. There was
also a check for ushell component when the service was requestet
synchronously. Now this ushell component check is added again.

Change-Id: I6c9a3272df700d4a27e8136296ab15ea577cfd0f