[INTERNAL] Table: Fix unstable tests for the Auto row mode with ODataV2
Because FPS can be lower in some test environments and the Auto mode uses
requestAnimationFrame, the order of events may not be as expected by some tests.

For example:
Expected: refreshRows -> async rendering -> updateRows
Actually: refreshRows -> updateRows -> async rendering

If updateRows is executed before the Auto mode has initially calculated the
number of rows and created them, contexts will not be requested.

To solve timing issues, the response time of the MockServer is increased from
10ms to 200ms to give the browser enough time to paint the next frame.

Change-Id: Ib1e49154260d7a90d3ea4c979ea2542bb6d88a0e