[FIX] sap.f.DynamicPage: Correct snap issue
Problem:
Extra whitespace bellow the header after snap. This happened because
upon snap, the page could not scroll to hide the entire header in
the scroll overflow because there is not sufficient scrollHeight.

Root cause:
The function _canSnapHeaderOnScroll wrongly returned 'true' for the
case when the header is in the title area. It used an extra check which
is redundant and even incorrect because when a header is moved from the
title area into the content area, the clientHeight of the scroll container
*automatically expands* => no need to additionally correct the maxScrollHeight.

Solution:
Remove the correction of the maxScrollHeight in _canSnapHeaderOnScroll

Change-Id: Ic32f72ec7e9f3a094d5d938c7ee8589c3e486285
BCP: 2170263509