[FIX] sap.m.Switch: fixed handle misplacement on drag in Quartz themes
* there was a misplacement of the handle of the switch when there are
no customTextOn/Off set (no labels), caused by the fact that in the
CSS parameters there are constraints for OFF and ON position, but the
switch without labels is shorter than one with labels in Quartz themes,
and the handle goes outside of the borders of the focus on drag operation.
In addition the width of the switch with no labels is different in Cozy
and Compact sizes, which cannot be covered by a single parameter.

* now this is fixed by adding a delta of the width as outline-offset CSS
parameter of the one of the inner containers, which effectively does not
affect the visualisation, but the delta can be set separately for Cozy
and Compact sizes, and later applied in the calculations.

* the fix does not affect other themes, since there is no outline-offset
set for the mentioned container. That way the delta is equal to 0 and
there is no correction at all.

Change-Id: Ibea8bfbaee62e3dfc16451c5102e314ef596dfac
CR-Id: 002075125800003440422021
BCP: 2170252080
(cherry picked from commit 82a402b694811471cf5852ba5b7358944edb5e70)