[FIX] sap.m.Select: Option with value state text is now announced
When SelectedKey of the Select and the ValueState or ValueStateText are dynamically changed
and bound to same property, there is a race condition of the announcement due to "Assertive"
behavior of the InvisibleMessage of the value state. When both the option and the valuestate is
changed, and the new value is "None", nothing is read.
With changing the Invisible Message mode to "polite" everything is read correctly.

BCP: 2170191779
Change-Id: I7a35ce71c5650f0207101bdf1bcf25cfea30ab36