[FIX] XMLView: Rendering of style attributes
Previous to this change the XMLView rendering produced invalid HTML
for nested <style> elements. The style elements textContent was
rendered directly.

If a <style> element is encountered during XML processing,
we need to remove the namespace prefix.

If a preprocessor is active we also need to make sure that the internal '__ui5'
namespace attribute is also not written to the DOM.

Change-Id: I45e32876c8113255a8ddc6caae3a37f4e30f5b61
BCP: 2170281098
BCP: 2170290345