[INTERNAL] sap.f.DynamicPage: Reverted fix footer overlap with content
This reverts commit 8b2c5ac4f0bddf7dedba5e6cc3e607bf07316ac2.

Reason for revert: sap.ui.table.Table does not respect max-height, which leads to content overlapping with the footer, no matter whether fitContent property is set to "true" or "false". As the usage of this table is most common scenario for the Dynamic Page, we need to revert this change and find another solution of the problem it fixes. It was done for not so common cases - Splitter in content or FCL. For the Splitter usage there is a work around - setting height not to 100%, but to 100% minus footer height. As for the FCL - it is not a supported scenario, Splitter can be used instead to visually separate the content.

Change-Id: I2e8e2310637637413b446bb705a194765e6c0c63
BCP: 2180333817