[INTERNAL] sap.m.Select: Change aria-roledescription to Listbox
The term "Combobox" means there are multiple ways to input a value, for example by typing it and choosing it.
With select the only way to input the value is by selecting it from a predefined options list. So base on user feedback the
value is changed to just "Listbox"

Change-Id: I623d4b1918fdc2008341e9c63f0c97b0cf05156f
CR-Id: 002075125800003687172021
BCP: 2180181715
(cherry picked from commit f70e89d86855167ff2e526f03dd98873fc9f0106)