[FIX] sap.m.DatePicker: Focus inside the picker when it is open
If we try to focus immediately after the calendar has been rendered,
the popup may not be positioned yet. This results in a page scroll
and bad picker position (not next to the input).

So we trigger the focus if the popup is already open or after it opens.

BCP: 2170299136
Change-Id: I0c2ade4f78333b1a3dae9190f11bca31f073c8f1