[INTERNAL] sap.m.Breadcrumbs: Correct focus behavior of select picker
Problem: There is additional tab stop for the sap.m.Select when width is not enough to render all of the items.
The focus navigates twice to the sap.m.Select.
The cause of this is that item navigation recieves the root dom of the items instead of their focus dom references.

Solution: Use the getFocusDomRef method when collecting the dom refs for the items for ItemNavigation.

Change-Id: Icc6a18056728703a8f1a0e7e282516437869fc8b
BCP: 2170255885