[FIX] sap.m.ObjectAttribute: 'ariaHasPopup' property now added
- A new property "ariaHasPopup" is now added to the
control public API. As a result if the control has active
text or a sap.m.Link in the "customContent" aggregation,
which opens a popover or dialog, it can be announced by
the screen readers.
- An active text or a sap.m.Link instance in the "customContent"
aggregation now become the active element when tabbed into.
- Accessibility information is moved from the root control element
to the active text or sap.m.Link element.
- Demokit samples improved in relation to the accessibility of the
active text.

BCP: 2170275459
Change-Id: I4b0eb26a43d41d5b2a0dbf7d43554fe1b1608b02