[FIX] sap.m.ProgressIndicator: Wrong visualization is fixed
When not set, flex child`s shrink and grow properties default values are 1. When all the
children
are with equal in these values, and one of them expands, the flex-basis of the smaller one is
ignored and it shrinks to fit along with the longer one.

Setting the shrink property to 0 makes the left part of the indicator not fall under its
dynamically set basis.

BCP: 2170289336
Change-Id: If351a4b2f93f1b7c5465b30f63afe20820596c38