[INTERNAL][FIX] sap.m.ProgressIndicator: Fixed wrong event bubbling
Problem:
When sap.m.ProgressIndicator is used inside sap.m.Table and is clicked,
the tap event is fired twice which leadс to execution of the handlers of
sap.m.ColumnListItem and sap.m.ProgressIndicator which
opens the information popover.

Fix:
Tap event is marked on its start phase as handled by sap.m.ProgressIndicator
the event bubbling is prevented.

BCP: 2170264495
Change-Id: Ic628cbc2ab3eda027febb3da355cb04d63c7caa1