[INTERNAL] sap.m.InputBase: Value State Announcement
JIRA: BGSOFUIRILA-3140

This change enables announcement of dynamic changes in the value state using a flag.
What is more with the change, the following was changed:
- Invisible Messaging is now introduced in InputBase and could be reused in other
use cases and controls.
- An announcement bug was found and fixed, related to the aria-live attribute.
Previously, the aria-live was set to the span referenced via aria-describedby/errormessage.
This resulted in a unneeded speech output, each time the value state is changed.
This is an issue, since we want to announce the value change, only when the control has the focus.
- Aria attributes are removed from the value state message, since it only has
presentational purposes and to avoid unwanted annoucement in future.

Change-Id: I9af5c77f3f4d9c7a704c68a63b09bdc4e6bf72ff