[FIX] sap.m.MultiComboBox: hasSelection now toggled correctly
The initial aim of the hasSelection property was to correctly toggle
the 'sapMMultiComboBoxHasToken' class. But after some issues,
the class toggle was moved inside a specific method, which was called
in the onAfterRendering delegate of the tokenizer. The property remained
but caused unwanted invalidation and issues.

However, the correct way to toggle the class is by changing the state, so all
the logic inside MultiComboBox.js, related to the class toggle was removed
and the hasSelection was used in a way, which correctly switches between states
and invalidated the control, when needed.

Since this leads to more rerendering, which is expected, some of the tests
where adapted or deleted.

Change-Id: I32d219e485a16f7c747423f453ac11b5604e16ab
BCP: 2170302260
CR-Id: 002075125800003898992021