[INTERNAL] sap.f.DynamicPageHeader: Remove aria-expanded attribute
Problem: Since aria 1.2 the aria-expanded attribute is not allowed on element with role region.
According to Fiori Screen Reader specification, if not allowed, the aria-expanded attribute should be replaced by aria-label atttribute with the expanded state value.

Solution: Remove the aria-expanded attribute and leave the aria-label to announce the expanded state of the header.

Change-Id: I52e6591f20a8cf1045fb03a984dc03af54ec6f79
BCP: 2170290368