[FIX] sap.uxap.ObjectPageLayout: Removed redundant accessibility readout
Problem: When we have sap.uxap.ObjectPageHeader as content of the
"headerTitle" aggregation, its "objectTitle" property is read twice
by the screenreader.

Root Cause: In ObjectPageLayout's "_getAriaLabelText" private function
we set the aria label for the root and for the header element to start
with the "objectTitle" property's value of the ObjectPageHeader.
This way, when we come from outside of the ObjectPageLayout and we land
on an element inside the ObjectPageHeader, the aria label of the
ObjectPageLayout is read followed by the aria label of the
ObjectPageHeader resulting in reading the "objectTitle" property's
value twice.

Solution: We no longer use "objectTitle" for the root element's
aria-label. Now the root element has a default aria-label and the
"objectTitle" is only passed on the aria label of the header element
thus it's read once.

BCP: 002028376700022310422021
Change-Id: I3d25d65e739246a66d76426f28d2e182090c88aa