[FIX] Demo Kit: Improved version selection user experience
Problem: When we select a major version (branch) from the Demo Kit's
version selection Dialog tree and click "Apply" we are redirected
to a non-existent URL.

Solution: We force the version group to point to the its latest version.
This way when we select a major version (branch) and click apply we are
redirected to its latest version without having to explicitly select it
from the branch's leaf.

Fixes: https://github.com/SAP/openui5/issues/3382

BCP: 2170314534
BCP: 2170315615
BCP: 2170317228
Change-Id: Id73b0ddb0a60eff6a98c7cf0d04a971467553a52