[INTERNAL] Table: Prevent reopening column menu on ENTER
The ColumnHeaderPopover closes immediately after activating a menu
button with Enter on keydown. The focus moves to the column header cell,
and the keyup event is triggered on the cell. The GridTable fires the
columnSelect event on keyup, which causes the MDC Table to open the menu
again.

A flag is set to prevent that the menu is opened again before being
fully closed.

Change-Id: Ie185899b318e8b718cda78b2e4d7c59c487bc305
BCP: 2170292560