[INTERNAL] Table: Select text in input elements of all types
Because the table sets the focus programmatically when navigating with
the keyboard, the browser does not select the text of input fields
automatically. The text is therefore also selected programmatically.

For cross-browser compatibility, the text had to be deselected when the
input lost the focus to avoid visual issues (text still appears
selected). This is no longer necessary in the latest browsers.

The browsers visually deselects the text on focus out. However,
selection read APIs still say that the text is selected (e.g.
selectionEnd > 0). But this is also the case for the browsers default
behavior and therefore acceptable.

Change-Id: I51037379650c56a2c0b96a5d36980156a1ad7729
BCP: 0020751294 0000761588 2021