[INTERNAL] ValueHelp: improve sample for DateTime
As the Service provides the dateTime in a UTC format, but the type parses it to some offset format we need to convert it.
As the service uses Milliseconds, but the user is not able to enter milliseconds we need to adjust the filter.
This is done in the table delegate.

Change-Id: I8f3c4c42d89e67abea9ede1c737a9277879010c8