[INTERNAL] Demo Kit: Fixed non-working links in sap.m.QuickView sample
Problem: When we click on a phone or an email link in the QuickView
popover, we are transferred to a page which states
"The requested resource was not found.".

Solution: We correctly parse the mail and phone targets in the
DocumentationRouter which results in working mail and phone links in
the QuickView sample.

BCP: 2170323317
Change-Id: Ic4eda6f703c762d68291498790783d32269df3b6