[FIX] sap.m.MenuButton: Now opens when enter pressed
Issue: We've had the events in different order, because our press is attached
to the sap.m.Button and the events came press -> keydown -> keyup.
Solution: Now we have predefined the current button onkeydown with adding
our logic to it. Our logic is needed because of acc reasons( we need to open
and close with enter when all options are disabled.

BCP: 2280154885
BCP: 2170321416
CR-Id: 002075125900001178002023
Change-Id: I77d4c98d9ba61393856af193a7d3da1a629ff29a
(cherry picked from commit 6f1c1ff995cb294068ae5c81e0590ee271436737)