[FIX] sap.m.MenuButton: Now opens when enter pressed
BCP: 2170321416

Issue: We've had the events in different order, because our press is attached
to the sap.m.Button and the events came press -> keydown -> keyup.
Solution: Now we have predefined the current button onkeydown with adding
our logic to it. Our logic is needed because of acc reasons( we need to open
and close with enter when all options are disabled.

Change-Id: I77d4c98d9ba61393856af193a7d3da1a629ff29a