[FIX] sap.uxap.HierarchicalSelect: Fixed lost of hierarchy
Problem: When navigating through the items in the popover via the
keyboard arrow keys, the hierarchy of the items is lost.

Root Cause: We don't apply the needed CSS styles when the list of the
HierarchicalSelect gets re-rendered.

Solution: Now every time the list with the items gets re-rendered, we
apply the hierarchy CSS styles.

BCP: 2170316918
Change-Id: Ia0128c990789e6dd097f33a6bd20c23b090af4ea