[INTERNAL] Demo Kit: Improved sap.f.FlexibleColumnLayout samples
Background: Some people reported that when navigating through items in
the tables presented in the FCL columns, the currently selected is not
marked as selected (blueish background).

Solution: We change the tables 'mode' property to 'SingleSelectMaster'.
This also resulted in some minor changes in the event handling when
selecting an item, in order to keep the previous item selection
behavior.

Additionally some indentation issues in the views are correctly aligned.

JIRA: BGSOFUIPIRIN-5457
Change-Id: I3ea0cec4b0d6b76e71a4654b1e51162bb96c31fb