[FIX] sap.m.FacetFilter: scroll arrows work
Problem:
The carousel with the lists has scroll arrows when it has many
lists. Clicking the arrows didn't do anything, because the arrow icons
didn't get detected properly as the click targets (due to a change in
the Icon DOM).
Solution:
The arrow icons get detected as the click targets, so the carousel
is scrolled.

BCP: 2170308086
Change-Id: I4be8f57109da1456805b57e7f20f536fd1787861