[INTERNAL]TwFB: improve valueHelp association in combination with p13n
The TableDelegate offers one hook to create a column and provides the
possibility to define a filter delegate in addition. As the column
might need a different fieldHelp associated to the inner Field template,
it might be necessary to associate a different fieldHelp to the FilterField
of the inner AdaptationFilterBar.

TableDelegate#addItem --> can be used to create a column
TableDelegate#getFilterDelegate#addItem --> can be used to create a FilterField

The example has ben enhanced to allow filtering also for "Orders" and "Authors".
In addition the same filter delegate as for the FilterBar delegate is being used,
this way the FilterFields will be created similarly as for the FilterBar.
For example "Stock" is now also a range slider.

Fixes testing issue #361

Change-Id: I3ac940f6b208f0b4c9c27212a47042c81668544b