[INTERNAL][FIX] sap/ui/fl can request base app flex/data for not yet deployed app variants
In cased a app variant is created within the IDE, the app variant itself may not yet be deployed in the back end system which serves the base application.
If this is the case, the flex/data request must be sent for the base application and also limit the layer.

Change-Id: I2a4d71f91be1c7ea50e6ecbb1aee88e5d56524ad
BCP: 2070117462
CR-Id: 002075125800001077402022
(cherry picked from commit 559a6aa58950b3421493f9401aaa84a2e732c945)