[FIX] sap.ui.model.analytics.AnalyticalBinding: no need to include automatically added columns to select
If there is a select parameter given it is no longer needed to add
the property names of columns that the AnalyticalBinding automatically
adds. For example the dimension property if only the text property is
given as a visible/inResult column. Before this fix, the properties
contained in the select binding parameter have been ignored in that case.

http://localhost:8080/resources/sap/ui/test/starter/Test.qunit.html?testsuite=test-resources%2Fsap%2Fui%2Fcore%2Fqunit%2Finternal%2Ftestsuite.models.qunit&test=AnalyticalBinding&hidepassed

PS1: Downport to rel-1.96

Change-Id: I10307d732806d861237c1b8fc6dccc716de628aa
BCP: 002075129400007616402021
CR-Id: 002075125900005366192021
(cherry picked from commit ec6a4bd4a6e59c00246036715931a5c3c62cc776)