[INTERNAL] FilterField: clear search value after closing typeahead
If user enters a character the typeahead opens. If multi-selection is enabled he can enter multiple items, the entered characters stay in the FilterField. If the popover is closed (user left the FilterField) the entered characters disapears as he has choosen values from typeahead. But if the user now opens the value help dialog, the old characters are still used in the search field.

Change-Id: I0c29734cccb8c653c7ce61c6264e14e7ac9593d1