[INTERNAL] sap.m.Dialog: Avoid sync request
Problem:
IconPool.getIconURI loads the resource bundle of sap/ui/core synchrounously,
if it's not available yet. This is the case when sap/m/Dialog is part of a custom bundle,
where the core is not yet initialized.

Solution:
Call IconPool.getIconURI later (in Dialog.prototype.setState),
rather than in the module definition of Dialog.

Snippix: 43263
Change-Id: I798203e8eb7724342e27029aeb8f06601032b879
BCP: 002075129400008328232021