[FIX] sap.uxap.ObjectPageLayout: Corrected check for closest section
Problem:
Undesired scroll to the *top* of the current section, because
the _isClosestScrolledSection failed to identify that the page
is already scrolled to the expected current section.
This is because the page was scrolled to a lower subSection
within the current section.

Solution:
Corrected the check in _isClosestScrolledSection to make provision
that the page may be scrolled to a lower subSection of the current section.

Renamed the function that identifies the current sectionBase to
_getClosestScrolledSectionBaseId to better indicate that it returns
a sectionBASE (section OR subSection) and not just section.

BCP: 2180393065
Change-Id: Ie48632f2dd0cc7ce755010a68718aed06f9f5c78