[INTERNAL] Demo Kit: Removed redundant resize listener from FCL sample
Problem: The user reports that we are accessing static
member (TABLET_BREAKPOINT) of the sap.f.FlexibleColumnLayout which in
not documented in one of our samples.

Solution: We are no longer accessing this static member. Its purpose
is to be used inside the FlexibleColumnLayout privately for its column
span calculations. Also the code which was accessing it in the sample
wasn't doing anything (dead code).

Fixes: https://github.com/SAP/openui5/issues/3404
BCP: 2180394948
Change-Id: If9f1a2d6048ad63939efc6ad6c43252337ebb873