[INTERNAL] sap.m.IconTabBar: Change role attribute
The overflow is of type IconTabFilter and had the role=tab attribute but as it is not part of the tablist a validation error occurred: An element with role=tab must be contained in, or owned by, an element with role=tablist. Now to fix this the overflow role is changed to "button".
NOW Number: CS20210002417845
SAP Number: 851323/2021

Change-Id: I7089d57d9ce6b909279a7b52f211149dddf55095
BCP: N/A