[INTERNAL] sap.m.IconTabBar: Change role attribute
The overflow is of type IconTabFilter and had the role=tab attribute but as it is not part of the tablist a validation error occurred: An element with role=tab must be contained in, or owned by, an element with role=tablist. Now to fix this the overflow role is changed to "button".
Change-Id: I7089d57d9ce6b909279a7b52f211149dddf55095
BCP: 002075129500008513232021
(cherry picked from commit 123a51df1b5a09b10496c02d95c12bfae3ce9dd0)
CR-Id: 002075125900005423542021