[FIX] sap.ui.core.format.DateFormat: Allow case-insensitive inputs
Allow case-insensitive inputs for: era, month, monthStandalone, quarter,
quarterStandalone and dayNameInWeek.

This is more user-friendly when parsing inputs such as for month (en):
May, may, MAY all result in the 5. month of the year: May

The string comparison is locale based that the characters are
correctly interpreted.

Accents are not ignored during the comparison, only the case, e.g.
a ≠ b, a ≠ á, a = A.

Store the locale of the loaded CLDR json file in
LocaleData#sCLDRLocaleId.

BCP: 2180419798
JIRA: CPOUI5FOUNDATION-459
Change-Id: Ic1e51d6b20e7498254fac1db5873cd185d2d0b8c