[FIX] sap.f.DynamicPage: Fix scroll issue with sticky subHeader
Problem: When there is a sticky subHeader provider set, the user is not
able to scroll down to last element in a Table, placed as content of
DynamicPage, when using the left button of the mouse.

Solution: It seems the height of the sticky subHeader is not got into
consideration, when calculating the contentSize of the Scrollbar. Now it
is added in the calculations.

BCP: 2170326966
Change-Id: If486a0d1e3d056f05144c15da06d40f311cff798