[INTERNAL] m.Table/ui.Table: Improve navigation from MessagePopover to Table
Actually, in most of the case, the links in the MessageDialog try to directly focus the given element that has reported an error.
Unfortunately it could be the case that this control, based on the page scrolled position, is even not visible in the screen at all.
If in a GridTable scenario the activeTitlePress event of the MessagePopover is implemented to focus the given element just by the calculated item index position
in the table, it happens that a completely different element gets the focus due of the fact that the original one is not visible on the screen.

Another example could be that the column, where an item has reported an error, is even not included via P13n in the table at all.
Where should the focus go in that case?

To avoid such scenarios the oFocusInfo in the focus event of core.Element object has been extended by an additional attribute.
The event itself gets than overwritten in the sap.m.Table and sap.ui.table.Table with table related logic.

JIRA: CPOUIFTEAMB-1714
Change-Id: I3bb3ab8c402e343575ec362518dee290df6450e4